refactor(settings): move recipe import into a Backup section under App Settings
Relocate the Import Recipes entry out of the top-level More screen and into a
new "Backup" section in App Settings, placed below the Browser section. Import
navigation now flows through AppSettingsBloc instead of the bottom-nav/Settings
path. This groups data-management actions together and leaves room for an
upcoming Export action in the same section.
