+++ client/bottomnav/public/src/commonMain/kotlin/com/plusmobileapps/chefmate/recipe/bottomnav/BottomNavBloc.kt
@@ -64,8 +64,6 @@ interface BottomNavBloc : BackHandlerOwner, BackClickBloc, BlocScreen {
 
         data object OpenAiChat : Output()
 
-        data object OpenImportRecipes : Output()
-
         data object OpenDeveloperSettings : Output()
 
         data class OpenUrl(val url: String) : Output()
