+++ client/settings/impl/src/commonMain/kotlin/com/plusmobileapps/chefmate/settings/impl/SettingsBlocImpl.kt
@@ -79,10 +79,6 @@ class SettingsBlocImpl(
         output.onNext(Output.OpenAiChat)
     }
 
-    override fun onImportRecipesClicked() {
-        output.onNext(Output.OpenImportRecipes)
-    }
-
     override fun onDeveloperSettingsClicked() {
         output.onNext(Output.OpenDeveloperSettings)
     }
