+++ client/settings/public/src/commonMain/composeResources/values/strings.xml
@@ -26,6 +26,7 @@
     <string name="app_settings_clear_history_dialog_cancel">Cancel</string>
     <string name="app_settings_navigation_section">Navigation</string>
     <string name="app_settings_bottom_nav_order">Bottom navigation order</string>
+    <string name="app_settings_backup_section">Backup</string>
     <string name="developer_settings">Developer Settings</string>
     <string name="version_label">Version {version}</string>
 </resources>
