+++ client/ui/screenshot-test/src/screenshotTestDebug/reference/com/plusmobileapps/chefmate/ui/screenshot/AppSettingsScreenshotTestKt/AppSettingsDarkScreenshot_4d30ee2f_0.png
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:3c0b59990619fd7f7dc04ba55bf0a352be3e3e7d6136a21237b89e82ad3a534a
-size 50204
+oid sha256:1a385a4b19f2b460f028be6d3b825ea8a6077091673aab907b89c5de45ac58b7
+size 57400
