+++ client/ui/screenshot-test/src/screenshotTestDebug/reference/com/plusmobileapps/chefmate/ui/screenshot/AppSettingsScreenshotTestKt/AppSettingsLightScreenshot_748aa731_0.png
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:b3b0b3fdb63de9e862bc452c0f2256252aacb9ef209cc2469b62c35995d1887d
-size 49739
+oid sha256:bc8cabac1cb7d01f349f527210908da17e27a09bc614a6aabefd2689f03fbc4d
+size 56728
